Start the webpack dev server bound to a specific DNS name instead of localhost
This lets me set the IP address of the given name in /etc/hosts,
which gets it to bind to the externally facing IP address of the VM
it's running in.

This means I can run the whole shebang in a VM, and don't have to
pollute my CentOS 7 workstation with node.